Move feature segment types to HotMFeatureSegmentTypes and initialize in register()
