+++ src/main/kotlin/com/github/hotm/world/gen/feature/HotMFeatures.kt
@@ -2,6 +2,7 @@ package com.github.hotm.world.gen.feature
 
 import com.github.hotm.HotMConstants
 import com.github.hotm.world.gen.feature.decorator.HotMDecorators
+import com.github.hotm.world.gen.feature.segment.HotMFeatureSegmentTypes
 import com.github.hotm.world.gen.feature.segment.SegmentedFeature
 import com.github.hotm.world.gen.feature.segment.SegmentedFeatureConfig
 import net.minecraft.util.registry.Registry
@@ -77,6 +78,7 @@ object HotMFeatures {
             register("nns_nectere_portal", NecterePortalFeature(DefaultFeatureConfig.CODEC))
 
         HotMDecorators.register()
+        HotMFeatureSegmentTypes.register()
 
         HotMStructureFeatures.register()
